:art:rename: api 함수 recipients,messages로 파일 분리
