+++ src/pages/ListPage.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { getList } from "apis/ListPage";
+import { getList } from "apis/messages";
 import Carousel from "../components/ListPage/Carousel";
 import "./ListPage.module.scss";
 import styles from "./ListPage.module.scss";
@@ -38,33 +38,39 @@ function ListPage() {
   return (
     <div className={styles.wrapper}>
       <div className={styles.track}>
-        <h1 className={`font-28-bold ${styles["title"]}`}>인기 롤링 페이퍼🔥</h1>
+        <h1 className={`font-28-bold ${styles["title"]}`}>
+          인기 롤링 페이퍼🔥
+        </h1>
         <div className={styles.carousel}>
-          <Carousel slideItems={bestItems} isLoading={isLoading}/>
+          <Carousel slideItems={bestItems} isLoading={isLoading} />
         </div>
         <div className={styles.empty}></div>
         <div className={`${styles["vertical-scroll"]} ${styles["item-1"]}`}>
           {bestItems?.map((item) => (
-            <CardList key={item.id} slideItems={item}/>
+            <CardList key={item.id} slideItems={item} />
           ))}
         </div>
-        <h1 className={`font-28-bold ${styles["title"]}`}>최근에 만든 롤링 페이퍼⭐</h1>
+        <h1 className={`font-28-bold ${styles["title"]}`}>
+          최근에 만든 롤링 페이퍼⭐
+        </h1>
         <div className={styles.carousel}>
-          <Carousel slideItems={recentItems} isLoading={isLoading}/>
+          <Carousel slideItems={recentItems} isLoading={isLoading} />
         </div>
         <div className={styles.empty}></div>
         <div className={`${styles["vertical-scroll"]} ${styles["item-2"]}`}>
           {recentItems?.map((item) => (
-            <CardList key={item.id} slideItems={item}/>
+            <CardList key={item.id} slideItems={item} />
           ))}
         </div>
       </div>
       <div className={`styles.["btn-box"]`}>
-          <Link
-            to="/post"
-            className={`button width-280 align-center font-18 ${styles["list-button"]}`}
-          >나도 만들어 보기</Link>
-          </div>
+        <Link
+          to="/post"
+          className={`button width-280 align-center font-18 ${styles["list-button"]}`}
+        >
+          나도 만들어 보기
+        </Link>
+      </div>
     </div>
   );
 }
